+++ packages/solana/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ant-design/web3-solana
 
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies [9bb1ea1]
+  - @ant-design/web3-common@1.13.0
+  - @ant-design/web3-assets@1.10.2
+
 ## 1.1.9
 
 ### Patch Changes